requirements.txt
anyio==4.12.1
certifi==2026.1.4
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.1
packaging==26.0
pillow==12.1.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==22.5
six==1.17.0
SQLAlchemy==2.0.46
typing_extensions==4.15.0
